@@ -3,25 +3,32 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "vue-cli-service serve --port 8080 --open",
-    "build": "vue-cli-service build"
+    "build": "vue-cli-service build",
+    "lint": "vue-cli-service lint",
+    "dev": "vue-cli-service serve --port 8080"
   },
   "dependencies": {
-    "axios": "^0.18.0",
-    "font-awesome": "4.7",
-    "nedb": "^1.8.0",
-    "vue": "^2.5.16",
-    "vue-lazyload": "^1.2.3",
-    "vue-router": "^3.0.1",
-    "vue-youtube": "^1.2.5",
-    "vuex": "^3.0.1"
+    "axios": "~0.18.1",
+    "boxicons": "^2.0.3",
+    "memory-cache": "^0.2.0",
+    "vue": "^2.6.10",
+    "vue-router": "^3.1.3",
+    "vue-youtube": "^1.3.5",
+    "vuex": "^3.1.1"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.0.0-beta.10",
-    "@vue/cli-service": "^3.0.0-beta.10",
-    "node-sass": "^4.7.2",
-    "sass-loader": "^6.0.6",
-    "vue-template-compiler": "^2.5.13"
+    "@vue/cli-plugin-babel": "^3.11.0",
+    "@vue/cli-plugin-eslint": "^3.11.0",
+    "@vue/cli-service": "^3.11.0",
+    "@vue/eslint-config-prettier": "^5.0.0",
+    "babel-eslint": "^10.0.1",
+    "eslint": "^5.16.0",
+    "eslint-plugin-prettier": "^3.1.0",
+    "eslint-plugin-vue": "^5.0.0",
+    "node-sass": "^4.12.0",
+    "prettier": "^1.18.2",
+    "sass-loader": "^8.0.0",
+    "vue-template-compiler": "^2.6.10"
   },
   "babel": {
     "presets": [
