@@ -1,21 +1,37 @@
 <template lang="html">
   <div class="star-rating">
-    <label class="star-rating__star" v-for="rating in ratings"
-    :class="{'is-selected': ((value >= rating) && value != null),'is-disabled': disabled}"
-    v-on:click="set(rating)" v-on:mouseover="star_over(rating)" v-on:mouseout="star_out">
-    <input class="star-rating star-rating__checkbox" type="radio" :value="rating" :name="name"
-    v-model="value" :disabled="disabled">★</label>
+    <label
+      class="star-rating__star"
+      v-for="(rating, index) in ratings"
+      :key="index"
+      :class="{
+        'is-selected': value >= rating && value != null,
+        'is-disabled': disabled
+      }"
+      @click="set(rating)"
+      @mouseover="star_over(rating)"
+      @mouseout="star_out"
+    >
+      <input
+        class="star-rating star-rating__checkbox"
+        type="radio"
+        :value="rating"
+        :name="name"
+        v-model="value"
+        :disabled="disabled"
+      />★</label
+    >
   </div>
 </template>
 
 <script>
 export default {
   props: {
-    'name': String,
-    'value': null,
-    'id': String,
-    'disabled': Boolean,
-    'required': Boolean
+    name: String,
+    value: null,
+    id: String,
+    disabled: Boolean,
+    required: Boolean
   },
   data: function() {
     return {
@@ -25,36 +41,27 @@ export default {
   },
 
   methods: {
-    star_over: function(index) {
-      var self = this;
-
+    star_over(index) {
       if (!this.disabled) {
         this.temp_value = this.value;
-        return this.value = index;
+        return (this.value = index);
       }
-
     },
 
-
-    star_out: function() {
-      var self = this;
-
+    star_out() {
       if (!this.disabled) {
-        return this.value = this.temp_value;
+        return (this.value = this.temp_value);
       }
     },
 
-
-    set: function(value) {
-      var self = this;
-
+    set(value) {
       if (!this.disabled) {
         this.temp_value = value;
-        return this.value = value;
+        return (this.value = value);
       }
     }
   }
-}
+};
 </script>
 
 <style lang="scss">
@@ -62,8 +69,11 @@ export default {
   position: absolute;
   overflow: hidden;
   clip: rect(0 0 0 0);
-  height: 1px; width: 1px;
-  margin: -1px; padding: 0; border: 0;
+  height: 1px;
+  width: 1px;
+  margin: -1px;
+  padding: 0;
+  border: 0;
 }
 
 .star-rating {
@@ -74,15 +84,15 @@ export default {
     vertical-align: middle;
     line-height: 1;
     font-size: 1.5em;
-    color: #ABABAB;
-    transition: color .2s ease-out;
+    color: #ababab;
+    transition: color 0.2s ease-out;
 
     &:hover {
       cursor: pointer;
     }
 
     &.is-selected {
-      color: #FFD700;
+      color: #ffd700;
     }
 
     &.is-disabled:hover {
@@ -94,5 +104,4 @@ export default {
     @extend %visually-hidden;
   }
 }
-
 </style>
