@@ -1,3 +1,3 @@
-export const shorten = (text,n)=> {
-  return (text.length > n) ? text.substr(0, n - 1) + '...' : text;
+export const shorten = (text, n) => {
+  return text.length > n ? text.substr(0, n - 1) + "..." : text;
 };
