@@ -1,18 +1,16 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
+import Vue from "vue";
+import Vuex from "vuex";
 
-Vue.use(Vuex)
+Vue.use(Vuex);
 
 export default new Vuex.Store({
   state: {
-    page:1
+    page: 1
   },
   mutations: {
-    setPage(state,page){
-      state.page=page;
+    setPage(state, page) {
+      state.page = page;
     }
   },
-  actions: {
-
-  }
-})
+  actions: {}
+});
